submitted version complete

+++ .Level2/level2.sh
@@ -64,6 +64,7 @@ player_y=2
 fi
 
 function displayFileMap(){
+clear
 for (( i=1; i<no_lines; i++ ));
 do
 echo
@@ -81,13 +82,16 @@ echo
 }
 function ifFonEqualN(){
 cat .level2Note
+read -n 1 -s -r -p "Press any key to continue"
 }
 function ifFonEqualC(){
 if test -f "Apple"; then
                         cat ./.noblePawnChat2
+						read -n 1 -s -r -p "Press any key to continue"
                 else
                         cat ./.noblePawnChat1
                         player_x=$(( player_x-1 ))
+						read -n 1 -s -r -p "Press any key to continue"
                 fi
 
 }
@@ -96,61 +100,38 @@ function ifFonEqualE(){
 	cp -r ../.Level3 ../Level3
         cat .pawnNote2
 	is_end=1
-	break
-}
-function checkChar(){
-	#C - character, N - note
-	if [[ "$fon" == "N" ]]; then
-		ifFonEqualN
-	elif [[ "$fon" == "C" ]]; then
-		ifFonEqualC
-	elif [[ "$fon" == "E" ]]; then
-		ifFonEqualE
-		is_end=1
-	fi
+	read -n 1 -s -r -p "Press any key to continue"
 }
 function moveLeft(){
+clear
 player_x2=$(( player_x-1 ))
 	fon=${file_map[$player_y,$player_x2]}
-	if [[ "$fon" == "|" ]] || [[ "$fon" == "X" ]]; then
-	echo "Ouch"
-	else
-	clear
-	player_x=$(( player_x -1 ))
-	displayFileMap
+	if [[ "$fon" != "|" ]] && [[ "$fon" != "X" ]]; then
+	player_x=$player_x2
 	fi
 }
 function moveRight(){
+clear
 	player_x2=$(( player_x+1 ))
         fon=${file_map[$player_y,$player_x2]}
-        if [[ "$fon" == "|" ]] || [[ "$fon" == "X" ]]; then
-	echo "Ouch"
-        else
-	clear
+		if [[ "$fon" != "|" ]] && [[ "$fon" != "X" ]]; then
         player_x=$player_x2
-        displayFileMap
         fi
 }
 function moveUp(){
+clear
  player_y2=$(( player_y-1 ))
         fon=${file_map[$player_y2,$player_x]}
-        if [[ "$fon" == "|" ]] || [[ "$fon" == "X" ]]; then
-	echo "Ouch"
-        else
-	clear
+		if [[ "$fon" != "|" ]] && [[ "$fon" != "X" ]]; then
         player_y=$player_y2
-        displayFileMap
         fi
 }
 function moveDown(){
+clear
         player_y2=$(( player_y+1 ))
         fon=${file_map[$player_y2,$player_x]}
-        if [[ "$fon" == "|" ]] || [[ "$fon" == "X" ]]; then
-	echo "Ouch!"
-        else
-	clear
+		if [[ "$fon" != "|" ]] && [[ "$fon" != "X" ]]; then
         player_y=$player_y2
-        displayFileMap
         fi
 }
 function rest(){
@@ -160,12 +141,17 @@ function rest(){
 	is_end=1
 }
 function reset(){
-echo 'temporary'
-rm .resume.txt
+if [ -f ".resume.txt" ] 
+then
+  rm .resume.txt
+  echo 'removed .resume.txt'
+fi
+is_end=1
 }
 function move(){
 is_moving=1
 while [[ $is_moving==1 ]]; do
+	displayFileMap
 	read -p "What direction? " -n 1 movement
 			case $movement in
 				"a")
@@ -190,11 +176,20 @@ while [[ $is_moving==1 ]]; do
 				;;
 			esac
 		fon=${file_map[$player_y,$player_x]}
-		checkChar
+			#C - character, N - note
+	if [[ "$fon" == "N" ]]; then
+		ifFonEqualN
+	elif [[ "$fon" == "C" ]]; then
+		ifFonEqualC
+	elif [[ "$fon" == "E" ]]; then
+		ifFonEqualE
+		break;
+	fi
 done
 }
-function whatAction(){
+function runGame(){
 while [[ $is_end == 0 ]]; do
+	displayFileMap
 	read -p "What Action? " action
 		if [[ "$action" == "rest" ]]; then
 			rest
@@ -205,15 +200,9 @@ while [[ $is_end == 0 ]]; do
 		fi
 done
 }
-function runGame(){
-while [[ $is_end == 0 ]]; do
-	whatAction
-done
-}
 
 function main(){ #an emulation of how the main function like that of C/C++ works. I will run everything in this
 setArrayContents
-displayFileMap
 runGame
 }
 main
